Add skip button as an icon
