+++ src/Layout/Header/Header.js
@@ -4,7 +4,6 @@ import { useIntl } from 'react-intl';
 /** Services */
 import http from '@services/http';
 import { login } from '@services/auth';
-import { getCurrent, skipCurrentSong } from '@services/api';
 /** Actions */
 import { setTokenAction, logoutAction } from '@context/auth';
 import { fetchMeAction } from '@context/me';
@@ -36,15 +35,6 @@ const Header = () => {
     dispatch(logoutAction());
   };
 
-  const skipHandler = async () => {
-    const { current } = await getCurrent();
-    console.log(current);
-    if (current) {
-      const roomSessionModel = { roomSessionId: current.publicId };
-      skipCurrentSong(roomSessionModel);
-    }
-  }
-
   useEffect(() => {
     if (window.location.hash) {
       const _token = retrieveSpotifyToken();
@@ -62,7 +52,7 @@ const Header = () => {
 
   return (
     <HeaderWrapper>
-      <HeaderLogo onClick={skipHandler}>Rogue Sound</HeaderLogo>
+      <HeaderLogo>Rogue Sound</HeaderLogo>
       <HeaderActionsWrapper>
         {token && <DeviceSelector intl={intl} />}
         {!token && (
